+++ ft_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: aoudija <aoudija@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/08 00:16:40 by aoudija           #+#    #+#             */
-/*   Updated: 2022/10/21 22:39:11 by aoudija          ###   ########.fr       */
+/*   Updated: 2022/10/23 09:17:29 by aoudija          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ void	*ft_memcpy(void *dst, const void *src, size_t n)
 {
 	char	*d;
 	char	*s;
-	int		i;
+	size_t	i;
 
 	i = 0;
 	d = (char *) dst;
